+++ CHANGELOG.md
@@ -13,10 +13,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Fixed an is where -ReplaceToast on Show-UDToast would not replace a toast (https://github.com/ironmansoftware/universal-dashboard/issues/1449) by [adamdriscoll](https://github.com/adamdriscoll)
+- Changed web.config to use PowerShell.exe rather than UD.Server.exe by [adamdriscoll](https://github.com/adamdriscoll)
 
 ### Changed (Enterprise)
 
 - Fixed issue where the login button wouldn't change font color by [adamdriscoll](https://github.com/adamdriscoll)
+
+### Removed 
+
+- Removed Publish-UDDashboard by [adamdriscoll](https://github.com/adamdriscoll)
+- Removed UniversalDashboard.Server.exe by [adamdriscoll](https://github.com/adamdriscoll)
 
 ## 2.8.3 - (1-31-2020)